+++ src/web/SeriesResource.java
@@ -6,16 +6,20 @@ import javax.ejb.EJBException;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
+import javax.ws.rs.CookieParam;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.Path;
+import javax.ws.rs.core.Cookie;
 import javax.ws.rs.core.Response;
 
 import comms.ResponseObject;
+import entity.Client;
 import entity.Series;
 
 import javax.ws.rs.PathParam;
 
 import service.SeriesService;
+import service.UserService;
 
 
 @Path("/series")
@@ -23,6 +27,9 @@ public class SeriesResource {
 
 	@EJB
 	SeriesService seriesService;
+	
+	@EJB
+	UserService userService;
 	
 	@GET
 	public Response getAllSeries() {
@@ -39,8 +46,12 @@ public class SeriesResource {
 	}
 	
 	@POST
-	public Response createSeries(Series s) {
+	public Response createSeries(@CookieParam("token") Cookie cookie, Series s) {
 		try{
+			Client c = userService.getClientByToken(cookie.getValue());
+			if(c.getAdmin() != 1){
+				return Response.status(Response.Status.UNAUTHORIZED).entity(new ResponseObject("Only admin can invoke this method!")).build();
+			}
 			seriesService.createSeries(s);
 			return Response.status(Response.Status.CREATED).entity(new ResponseObject("Series created successfully!")).build();
 		}
@@ -50,8 +61,12 @@ public class SeriesResource {
 	}
 	
 	@PUT
-	public Response editSeries(Series s) {
+	public Response editSeries(@CookieParam("token") Cookie cookie, Series s) {
 		try{
+			Client c = userService.getClientByToken(cookie.getValue());
+			if(c.getAdmin() != 1){
+				return Response.status(Response.Status.UNAUTHORIZED).entity(new ResponseObject("Only admin can invoke this method!")).build();
+			}
 			seriesService.editSeries(s);
 			return Response.status(Response.Status.OK).entity(new ResponseObject("Series updated successfully!")).build();
 		}
@@ -61,8 +76,12 @@ public class SeriesResource {
 	}
 	
 	@DELETE
-	public Response removeSeries(Series s) {
+	public Response removeSeries(@CookieParam("token") Cookie cookie, Series s) {
 		try{
+			Client c = userService.getClientByToken(cookie.getValue());
+			if(c.getAdmin() != 1){
+				return Response.status(Response.Status.UNAUTHORIZED).entity(new ResponseObject("Only admin can invoke this method!")).build();
+			}
 			seriesService.removeSeries(s);
 			return Response.status(Response.Status.OK).entity(new ResponseObject("Series removed successfully!")).build();
 		}
